Fixed some detekt issues
