+++ konstraints-core/src/main/kotlin/tools/aqua/konstraints/parser/ProtoCommand.kt
@@ -21,6 +21,7 @@ package tools.aqua.konstraints.parser
 import org.petitparser.context.Token
 import tools.aqua.konstraints.smt.*
 
+/** Symbol wrapper holding additional information from parser. */
 class ParseSymbol(val token: Token) :
     Symbol(
         token.getValue(),
