+++ konstraints-core/src/main/kotlin/tools/aqua/konstraints/smt/Expression.kt
@@ -21,7 +21,7 @@ package tools.aqua.konstraints.smt
 import tools.aqua.konstraints.parser.IteDecl
 import tools.aqua.konstraints.util.reduceOrDefault
 
-/** Interface for all sorted SMT terms */
+/** Interface for all sorted SMT terms. */
 sealed interface Expression<out T : Sort> {
   val name: SMTSerializable
   val sort: T
@@ -66,7 +66,7 @@ sealed interface Expression<out T : Sort> {
         is AnnotatedExpression -> predicate(this) and term.all(predicate)
       }
 
-  /** Convert [this] to a sequence starting with [this] followed by [children] */
+  /** Convert [this] to a sequence starting with [this] followed by [children]. */
   fun asSequence(): Sequence<Expression<*>> = sequence {
     yield(this@Expression)
     children.forEach { yieldAll(it.asSequence()) }
@@ -91,12 +91,13 @@ sealed interface Expression<out T : Sort> {
     }
   }
 
+  /** Create a copy of [this] using the provided [children]. */
   fun copy(children: List<Expression<*>>): Expression<T>
 
   val children: List<Expression<*>>
 }
 
-/** SMT Literal */
+/** SMT Literal. */
 abstract class Literal<out T : Sort>(override val name: LiteralString, override val sort: T) :
     Expression<T> {
   override val func = null
@@ -106,6 +107,7 @@ abstract class Literal<out T : Sort>(override val name: LiteralString, override 
   override fun toString() = name.toString()
 }
 
+/** Base class for all expressions with no children. */
 abstract class ConstantExpression<out T : Sort>(override val name: Symbol, override val sort: T) :
     Expression<T> {
   override val func = null
@@ -115,7 +117,7 @@ abstract class ConstantExpression<out T : Sort>(override val name: Symbol, overr
   override fun toString() = "$name"
 }
 
-/** Base class of all expressions with exactly one child */
+/** Base class of all expressions with exactly one child. */
 abstract class UnaryExpression<out T : Sort, out S : Sort>(
     override val name: Symbol,
     override val sort: T
@@ -130,7 +132,7 @@ abstract class UnaryExpression<out T : Sort, out S : Sort>(
   override fun toString() = "($name ${inner})"
 }
 
-/** Base class of all expressions with exactly two children */
+/** Base class of all expressions with exactly two children. */
 abstract class BinaryExpression<out T : Sort, out S1 : Sort, out S2 : Sort>(
     override val name: Symbol,
     override val sort: T
@@ -147,7 +149,7 @@ abstract class BinaryExpression<out T : Sort, out S1 : Sort, out S2 : Sort>(
   override fun toString() = "($name $lhs $rhs)"
 }
 
-/** Base class of all expressions with exactly three children */
+/** Base class of all expressions with exactly three children. */
 abstract class TernaryExpression<out T : Sort, out S1 : Sort, out S2 : Sort, out S3 : Sort>(
     override val name: Symbol,
     override val sort: T
@@ -168,7 +170,7 @@ abstract class TernaryExpression<out T : Sort, out S1 : Sort, out S2 : Sort, out
 
 /**
  * Base class of all expressions with any children, where all children are expressions of the same
- * sort
+ * sort.
  */
 abstract class HomogenousExpression<out T : Sort, out S : Sort>(
     override val name: Symbol,
@@ -183,7 +185,7 @@ abstract class HomogenousExpression<out T : Sort, out S : Sort>(
 }
 
 /**
- * Implements ite according to Core theory (par (A) (ite Bool A A A))
+ * Implements ite according to Core theory (par (A) (ite Bool A A A)).
  *
  * @param statement indicates whether [then] or [otherwise] should be returned
  * @param then value to be returned if [statement] is true
@@ -212,7 +214,7 @@ class Ite<out T : Sort>(
   override fun toString(): String = "(ite $statement $then $otherwise)"
 }
 
-/** Base class of all expressions with any number of children */
+/** Base class of all expressions with any number of children. */
 abstract class NAryExpression<out T : Sort>(override val name: Symbol, override val sort: T) :
     Expression<T> {
 
@@ -221,7 +223,7 @@ abstract class NAryExpression<out T : Sort>(override val name: Symbol, override 
       else name.toSMTString(QuotingRule.SAME_AS_INPUT)
 }
 
-/** Let expression */
+/** Let expression. */
 class LetExpression<out T : Sort>(val bindings: List<VarBinding<*>>, val inner: Expression<T>) :
     Expression<T> {
   override val sort = inner.sort
@@ -261,7 +263,7 @@ class LetExpression<out T : Sort>(val bindings: List<VarBinding<*>>, val inner: 
   override val children: List<Expression<*>> = listOf(inner)
 }
 
-/** Any expression generated by a user declared smt function */
+/** Any expression generated by a user declared smt function. */
 class UserDeclaredExpression<out T : Sort>(
     name: Symbol,
     sort: T,
@@ -278,7 +280,7 @@ class UserDeclaredExpression<out T : Sort>(
       UserDeclaredExpression(name, sort, children, func)
 }
 
-/** Any expression defined by a user defined smt function */
+/** Any expression defined by a user defined smt function. */
 class UserDefinedExpression<T : Sort>(
     name: Symbol,
     sort: T,
@@ -301,7 +303,7 @@ class UserDefinedExpression<T : Sort>(
   fun expand(): Expression<*> = definition.expand(children)
 }
 
-/** Expression with a local variable created via let */
+/** Expression with a local variable created via let. */
 class LocalExpression<T : Sort>(
     override val name: Symbol,
     override val sort: T,
@@ -321,7 +323,7 @@ class LocalExpression<T : Sort>(
   override val children: List<Expression<*>> = emptyList()
 }
 
-/** Existence quantified expression with variables [vars] */
+/** Existence quantified expression with variables [vars]. */
 class ExistsExpression(val vars: List<SortedVar<*>>, val term: Expression<BoolSort>) :
     Expression<BoolSort> {
   override val theories = emptySet<Theories>()
@@ -340,7 +342,7 @@ class ExistsExpression(val vars: List<SortedVar<*>>, val term: Expression<BoolSo
   }
 }
 
-/** Forall quantified expression with variables [vars] */
+/** Forall quantified expression with variables [vars]. */
 class ForallExpression(val vars: List<SortedVar<*>>, val term: Expression<BoolSort>) :
     Expression<BoolSort> {
   override val theories = emptySet<Theories>()
@@ -357,7 +359,7 @@ class ForallExpression(val vars: List<SortedVar<*>>, val term: Expression<BoolSo
   }
 }
 
-/** Instance of a bound variable inside exists or forall */
+/** Instance of a bound variable inside exists or forall. */
 class BoundVariable<out T : Sort>(
     override val name: Symbol,
     override val sort: T,
@@ -370,7 +372,7 @@ class BoundVariable<out T : Sort>(
   override fun copy(children: List<Expression<*>>): Expression<T> = BoundVariable(name, sort, func)
 }
 
-/** Expression with an Annotation */
+/** Expression with an Annotation. */
 class AnnotatedExpression<T : Sort>(val term: Expression<T>, val annoations: List<Attribute>) :
     Expression<T> {
   override val theories = emptySet<Theories>()
@@ -388,7 +390,7 @@ class AnnotatedExpression<T : Sort>(val term: Expression<T>, val annoations: Lis
   override val children: List<Expression<*>> = listOf(term)
 }
 
-/** Specialization of [ClassCastException], thrown when casting an expression to an illegal sort */
+/** Specialization of [ClassCastException], thrown when casting an expression to an illegal sort. */
 class ExpressionCastException(msg: String) : ClassCastException(msg)
 
 /**
