+++ konstraints-core/src/main/kotlin/tools/aqua/konstraints/smt/Ints.kt
@@ -21,9 +21,9 @@ package tools.aqua.konstraints.smt
 import tools.aqua.konstraints.parser.*
 
 /**
- * Integer negation
- *
- * (- Int Int)
+ * Integer negation.
+ * - (- Int Int)
+ * - (- [inner])
  */
 class IntNeg(override val inner: Expression<IntSort>) :
     UnaryExpression<IntSort, IntSort>("-".toSymbolWithQuotes(), SMTInt) {
@@ -34,9 +34,9 @@ class IntNeg(override val inner: Expression<IntSort>) :
 }
 
 /**
- * Integer subtraction
- *
- * (- Int Int Int :left-assoc)
+ * Integer subtraction.
+ * - (- Int Int Int :left-assoc)
+ * - (- [terms])
  */
 class IntSub(val terms: List<Expression<IntSort>>) :
     HomogenousExpression<IntSort, IntSort>("-".toSymbolWithQuotes(), SMTInt) {
@@ -57,9 +57,9 @@ class IntSub(val terms: List<Expression<IntSort>>) :
 }
 
 /**
- * Integer addition
- *
- * (+ Int Int Int :left-assoc)
+ * Integer addition.
+ * - (+ Int Int Int :left-assoc)
+ * - (+ [terms])
  */
 class IntAdd(val terms: List<Expression<IntSort>>) :
     HomogenousExpression<IntSort, IntSort>("+".toSymbolWithQuotes(), SMTInt) {
@@ -80,9 +80,9 @@ class IntAdd(val terms: List<Expression<IntSort>>) :
 }
 
 /**
- * Integer multiplication
- *
- * (* Int Int Int :left-assoc)
+ * Integer multiplication.
+ * - (* Int Int Int :left-assoc)
+ * - (* [factors])
  */
 class IntMul(val factors: List<Expression<IntSort>>) :
     HomogenousExpression<IntSort, IntSort>("*".toSymbolWithQuotes(), SMTInt) {
@@ -103,9 +103,9 @@ class IntMul(val factors: List<Expression<IntSort>>) :
 }
 
 /**
- * Integer division
- *
- * (div Int Int Int :left-assoc)
+ * Integer division.
+ * - (div Int Int Int :left-assoc)
+ * - (div [terms])
  */
 class IntDiv(val terms: List<Expression<IntSort>>) :
     HomogenousExpression<IntSort, IntSort>("/".toSymbolWithQuotes(), SMTInt) {
@@ -126,9 +126,9 @@ class IntDiv(val terms: List<Expression<IntSort>>) :
 }
 
 /**
- * Modulo
- *
- * (mod Int Int Int)
+ * Modulo.
+ * - (mod Int Int Int)
+ * - (mod [dividend] [divisor])
  */
 class Mod(val dividend: Expression<IntSort>, val divisor: Expression<IntSort>) :
     BinaryExpression<IntSort, IntSort, IntSort>("mod".toSymbolWithQuotes(), SMTInt) {
@@ -143,9 +143,9 @@ class Mod(val dividend: Expression<IntSort>, val divisor: Expression<IntSort>) :
 }
 
 /**
- * Absolute value
- *
- * (abs Int Int)
+ * Absolute value.
+ * - (abs Int Int)
+ * - (abs [inner])
  */
 class Abs(override val inner: Expression<IntSort>) :
     UnaryExpression<IntSort, IntSort>("abs".toSymbolWithQuotes(), SMTInt) {
@@ -156,9 +156,9 @@ class Abs(override val inner: Expression<IntSort>) :
 }
 
 /**
- * Integer less equals
- *
- * (<= Int Int Bool :chainable)
+ * Integer less equals.
+ * - (<= Int Int Bool :chainable)
+ * - (<= [terms])
  */
 class IntLessEq(val terms: List<Expression<IntSort>>) :
     HomogenousExpression<BoolSort, IntSort>("<=".toSymbolWithQuotes(), Bool) {
@@ -179,9 +179,9 @@ class IntLessEq(val terms: List<Expression<IntSort>>) :
 }
 
 /**
- * Integer less
- *
- * (< Int Int Bool :chainable)
+ * Integer less.
+ * - (< Int Int Bool :chainable)
+ * - (< [terms])
  */
 class IntLess(val terms: List<Expression<IntSort>>) :
     HomogenousExpression<BoolSort, IntSort>("<".toSymbolWithQuotes(), Bool) {
@@ -202,9 +202,9 @@ class IntLess(val terms: List<Expression<IntSort>>) :
 }
 
 /**
- * Integer greater equals
- *
- * (>= Int Int Bool :chainable)
+ * Integer greater equals.
+ * - (>= Int Int Bool :chainable)
+ * - (>= [terms])
  */
 class IntGreaterEq(val terms: List<Expression<IntSort>>) :
     HomogenousExpression<BoolSort, IntSort>(">=".toSymbolWithQuotes(), Bool) {
@@ -225,9 +225,9 @@ class IntGreaterEq(val terms: List<Expression<IntSort>>) :
 }
 
 /**
- * Integer greater
- *
- * (> Int Int Bool :chainable)
+ * Integer greater.
+ * - (> Int Int Bool :chainable)
+ * - (> [terms])
  */
 class IntGreater(val terms: List<Expression<IntSort>>) :
     HomogenousExpression<BoolSort, IntSort>(">".toSymbolWithQuotes(), Bool) {
@@ -248,11 +248,11 @@ class IntGreater(val terms: List<Expression<IntSort>>) :
 }
 
 /**
- * Divisible predicate
+ * Divisible predicate.
  *
  * @throws IllegalArgumentException if [n] <= 0
- *
- * ((_ divisible n) Int Bool)
+ * - ((_ divisible n) Int Bool)
+ * - ((_ divisible [n]) [inner])
  */
 class Divisible(val n: Int, override val inner: Expression<IntSort>) :
     UnaryExpression<BoolSort, IntSort>("divisible".toSymbolWithQuotes(), Bool) {
