+++ konstraints-core/src/main/kotlin/tools/aqua/konstraints/smt/Literal.kt
@@ -73,11 +73,9 @@ private constructor(vector: String, val bits: Int, val isBinary: Boolean, val va
 }
 
 /**
- * Floating-point literal
+ * Floating-point literal.
  *
- * @param sign bitvector of length 1, holding the sign
- * @param exponent bitvector holding the exponent value
- * @param significand bitvector holding the significand value
+ * (fp [sign] [exponent] [significand])
  */
 data class FPLiteral(
     val sign: Expression<BVSort>,
@@ -123,7 +121,7 @@ data class FPLiteral(
 }
 
 /**
- * Integer literals
+ * Integer literals.
  *
  * (NUMERAL Int)
  */
@@ -145,7 +143,7 @@ class IntLiteral(val value: BigInteger) :
 }
 
 /**
- * Real literal
+ * Real literal.
  *
  * (NUMERAL Real) (DECIMAL Real)
  */
@@ -177,7 +175,7 @@ class RealLiteral(val value: BigDecimal) :
 /**
  * [value] holds the hexadecimal unicode of the character, [character] hold the kotlin
  * representation, possible values range from #x0 to #x2FFFF and are generated by the following
- * grammar
+ * grammar.
  *
  * ```
  * ⟨H⟩ ::= #x⟨F⟩ | #x⟨F⟩⟨F⟩ | #x⟨F⟩⟨F⟩⟨F⟩ | #x⟨F⟩⟨F⟩⟨F⟩⟨F⟩ | #x⟨2⟩⟨F⟩⟨F⟩⟨F⟩⟨F⟩
