+++ konstraints-core/src/main/kotlin/tools/aqua/konstraints/smt/Reals.kt
@@ -21,9 +21,9 @@ package tools.aqua.konstraints.smt
 import tools.aqua.konstraints.parser.*
 
 /**
- * Real negation
- *
- * (- Real Real)
+ * Real negation.
+ * - (- Real Real)
+ * - (- [inner])
  */
 class RealNeg(override val inner: Expression<RealSort>) :
     UnaryExpression<RealSort, RealSort>("-".toSymbolWithQuotes(), Real) {
@@ -34,9 +34,9 @@ class RealNeg(override val inner: Expression<RealSort>) :
 }
 
 /**
- * Real subtraction
- *
- * (- Real Real Real :left-assoc)
+ * Real subtraction.
+ * - (- Real Real Real :left-assoc)
+ * - (- [terms])
  */
 class RealSub(val terms: List<Expression<RealSort>>) :
     HomogenousExpression<RealSort, RealSort>("-".toSymbolWithQuotes(), Real) {
@@ -57,9 +57,9 @@ class RealSub(val terms: List<Expression<RealSort>>) :
 }
 
 /**
- * Real addition
- *
- * (+ Real Real Real :left-assoc)
+ * Real addition.
+ * - (+ Real Real Real :left-assoc)
+ * - (+ [terms])
  */
 class RealAdd(val terms: List<Expression<RealSort>>) :
     HomogenousExpression<RealSort, RealSort>("+".toSymbolWithQuotes(), Real) {
@@ -80,9 +80,9 @@ class RealAdd(val terms: List<Expression<RealSort>>) :
 }
 
 /**
- * Real multiplication
- *
- * (* Real Real Real :left-assoc)
+ * Real multiplication.
+ * - (* Real Real Real :left-assoc)
+ * - (* [factors])
  */
 class RealMul(val factors: List<Expression<RealSort>>) :
     HomogenousExpression<RealSort, RealSort>("*".toSymbolWithQuotes(), Real) {
@@ -103,9 +103,9 @@ class RealMul(val factors: List<Expression<RealSort>>) :
 }
 
 /**
- * Real division
- *
- * (/ Real Real Real :left-assoc)
+ * Real division.
+ * - (/ Real Real Real :left-assoc)
+ * - (/ [terms])
  */
 class RealDiv(val terms: List<Expression<RealSort>>) :
     HomogenousExpression<RealSort, RealSort>("/".toSymbolWithQuotes(), Real) {
@@ -126,9 +126,9 @@ class RealDiv(val terms: List<Expression<RealSort>>) :
 }
 
 /**
- * Real less equals
- *
- * (<= Real Real Bool :chainable)
+ * Real less equals.
+ * - (<= Real Real Bool :chainable)
+ * - (<= [terms])
  */
 class RealLessEq(val terms: List<Expression<RealSort>>) :
     HomogenousExpression<BoolSort, RealSort>("<=".toSymbolWithQuotes(), Bool) {
@@ -149,9 +149,9 @@ class RealLessEq(val terms: List<Expression<RealSort>>) :
 }
 
 /**
- * Real less
- *
- * (<= Real Real Bool :chainable)
+ * Real less.
+ * - (<= Real Real Bool :chainable)
+ * - (<= [terms])
  */
 class RealLess(val terms: List<Expression<RealSort>>) :
     HomogenousExpression<BoolSort, RealSort>("<".toSymbolWithQuotes(), Bool) {
@@ -172,9 +172,9 @@ class RealLess(val terms: List<Expression<RealSort>>) :
 }
 
 /**
- * Real greater equals
- *
- * (<= Real Real Bool :chainable)
+ * Real greater equals.
+ * - (<= Real Real Bool :chainable)
+ * - (<= [terms])
  */
 class RealGreaterEq(val terms: List<Expression<RealSort>>) :
     HomogenousExpression<BoolSort, RealSort>(">=".toSymbolWithQuotes(), Bool) {
@@ -195,9 +195,9 @@ class RealGreaterEq(val terms: List<Expression<RealSort>>) :
 }
 
 /**
- * Real greater
- *
- * (<= Real Real Bool :chainable)
+ * Real greater.
+ * - (<= Real Real Bool :chainable)
+ * - (<= [terms])
  */
 class RealGreater(val terms: List<Expression<RealSort>>) :
     HomogenousExpression<BoolSort, RealSort>(">".toSymbolWithQuotes(), Bool) {
