+++ konstraints-core/src/main/kotlin/tools/aqua/konstraints/smt/SExpression.kt
@@ -22,20 +22,20 @@ import org.petitparser.context.Token
 
 // S-Expression
 
-/** Base class for explicit SExpressions */
+/** Base class for explicit SExpressions. */
 sealed interface SExpression
 
-/** SubSExpression */
+/** SubSExpression. */
 data class SubSExpression(val subExpressions: List<SExpression>) : SExpression
 
-/** SExpressionConstant */
+/** SExpressionConstant. */
 data class SExpressionConstant(val constant: SpecConstant) : SExpression
 
-/** SExpressionSymbol */
+/** SExpressionSymbol. */
 data class SExpressionSymbol(val symbol: Symbol) : SExpression
 
-/** SExpressionReserved */
+/** SExpressionReserved. */
 data class SExpressionReserved(val reserved: Token) : SExpression
 
-/** SExpressionKeyword */
+/** SExpressionKeyword. */
 data class SExpressionKeyword(val keyword: Token) : SExpression
