+++ konstraints-core/src/main/kotlin/tools/aqua/konstraints/smt/SpecConstant.kt
@@ -21,16 +21,16 @@ package tools.aqua.konstraints.smt
 import java.math.BigDecimal
 import java.math.BigInteger
 
-/** Base class for all smt constant types used in [AttributeValue] */
+/** Base class for all smt constant types used in [AttributeValue]. */
 sealed interface SpecConstant
 
-/** SMT String constant */
+/** SMT String constant. */
 data class StringConstant(val string: String) : SpecConstant
 
-/** SMT Numeral constant */
+/** SMT Numeral constant. */
 data class NumeralConstant(val numeral: BigInteger) : SpecConstant
 
-/** BinaryConstant of the form #b followed by a non-empty sequence of 0 and 1 characters */
+/** BinaryConstant of the form #b followed by a non-empty sequence of 0 and 1 characters. */
 data class BinaryConstant(val binary: String) : SpecConstant {
   /* Number of bits for this binary */
   val bits = binary.length - 2
@@ -38,12 +38,12 @@ data class BinaryConstant(val binary: String) : SpecConstant {
 
 /**
  * Hexadecimal constant of the form #x followed by a non-empty sequence of digits and letters from A
- * to F , capitalized or not
+ * to F , capitalized or not.
  */
 data class HexConstant(val hexadecimal: String) : SpecConstant {
   /* Number of bits for this hexadecimal */
   val bits = (hexadecimal.length - 2) * 4
 }
 
-/** Decimal smt constant */
+/** Decimal smt constant. */
 data class DecimalConstant(val decimal: BigDecimal) : SpecConstant
