+++ konstraints-core/src/main/kotlin/tools/aqua/konstraints/solvers/Solver.kt
@@ -23,16 +23,16 @@ import tools.aqua.konstraints.smt.SMTProgram
 import tools.aqua.konstraints.smt.SatStatus
 
 interface Solver : AutoCloseable {
-  /** Solves the provided program using this solver */
+  /** Solves the provided program using this solver. */
   fun solve(program: SMTProgram): SatStatus
 
-  /** A model if available, else null */
+  /** A model if available, else null. */
   val modelOrNull: Model?
 
-  /** The model if one was generated */
+  /** The model if one was generated. */
   val model: Model
     get() = checkNotNull(modelOrNull) { "model is null" }
 
-  /** True if a model is available */
+  /** True if a model is available. */
   val isModelAvailable: Boolean
 }
