+++ konstraints-core/src/main/kotlin/tools/aqua/konstraints/util/SimpleDelegate.kt
@@ -21,7 +21,7 @@ package tools.aqua.konstraints.util
 import kotlin.properties.ReadOnlyProperty
 import kotlin.reflect.KProperty
 
-/** Delegate that holds a single cached value */
+/** Delegate that holds a single cached value. */
 class SimpleDelegate<T>(val value: T) : ReadOnlyProperty<Any?, T> {
   override fun getValue(thisRef: Any?, property: KProperty<*>) = value
 }
